Rock_Paper_Scissors
Interactive game called Rock_Paper_Scissors. 
This game is made little bit more interactive with the users, find below the steps.

Game asks the names of the users. name 1 and name 2.
The users pick an option: 'Rock'= 1,'Scissors' = 2,'Paper'= 3
The Logic compare and provide the winner.
if the user pick the same answer no one is the winner.
the program ask to keep playing or stop the game.
if the aswer is keep playing, the program will ask to pick an option again.
however if the answer is not, the game stop and gives thanks to the players.